Clean up the dependencies
Problems were found and corrected thanks to the command:
    mvn dependency:analyze

It is an anti-pattern to have additional dependencies declared in an
active-by-default profile. And we can use version properties for the
development profile to switch desired dependencies to SNAPSHOT versions.

We should not use the AppleJavaExtensions because that JAR is not
legally redistributable; instead, we use the "orange" extensions,
which offer the same stubs, but without legal problems.